test: waydraymath and weth-gateway
